Cleaned up structure of firebase functions for stripe
Created utility 'onCall' function to streamline future firebase function creation.
Added the stripe payment intent id and amount charged into acuity booking.
Wrote readme explaining how the entire holiday program solution works.
