+++ firebase/functions/fizz-kidz/src/acuity/constants/formFields.ts
@@ -21,7 +21,9 @@ const FormFields = {
     CONTINUING_WITH_TERM: 9878436,
     CONTINUING_WITH_TERM_EMAIL_SENT: 9965760,
     EMERGENCY_CONTACT_NAME_HP: 11901970,
-    EMERGENCY_CONTACT_NUMBER_HP: 11901971
+    EMERGENCY_CONTACT_NUMBER_HP: 11901971,
+    HOLIDAY_PROGRAM_PAYMENT_INTENT_ID: 11929615,
+    HOLIDAY_PROGRAM_AMOUNT_CHARGED: 11929616
 }
 
 export default FormFields
