+++ firebase/functions/fizz-kidz/src/acuity/constants/forms.ts
@@ -8,7 +8,8 @@ const Forms = {
     INVOICE: 1337930,
     TERMS_AND_CONDITIONS: 1448907,
     CONTINUING_WITH_TERM: 1779820,
-    HOLIDAY_PROGRAM_EMERGENCY_CONTACT: 2132257
+    HOLIDAY_PROGRAM_EMERGENCY_CONTACT: 2132257,
+    HOLIDAY_PROGRAM_PAYMENT_DETAILS: 2136574
 }
 
 export default Forms
